--- src/main/java/roomescape/auth/AdminFilter.java
@@ -1,38 +1,0 @@
-package roomescape.auth;
-
-import javax.servlet.*;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class AdminFilter implements Filter {
-    JwtTokenProvider jwtTokenProvider;
-
-    public AdminFilter(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @Override
-    public void init(FilterConfig filterConfig) throws ServletException {
-        Filter.super.init(filterConfig);
-    }
-
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) {
-        try {
-            HttpServletRequest httpServletRequest = (HttpServletRequest) request;
-            String credential = httpServletRequest.getHeader("Authorization").split(" ")[1];
-            if (!jwtTokenProvider.validateToken(credential) || !jwtTokenProvider.getRole(credential).equals("ADMIN")) {
-                throw new AuthenticationException();
-            }
-            chain.doFilter(request, response);
-        } catch (Exception e) {
-            HttpServletResponse httpResponse = (HttpServletResponse) response;
-            httpResponse.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        }
-    }
-
-    @Override
-    public void destroy() {
-        Filter.super.destroy();
-    }
-}
